feat : 구글 Script(구글 애널리틱스) 컴퓨넌트화
